requirements.txt
streamlit==1.31.0
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.1
huggingface-hub>=0.34.0
plotly==5.18.0
pandas==2.1.4
